refactor(store): move the project-dir key into a leaf module
The schema bootstrap needs this to migrate old rows, and it was reaching for it
with a dynamic import of transcript-index - which imports database, which
imports the connection pool, which imports the bootstrap. A cycle resolved at
call time inside the one function every database open runs through.

Whether or not that produced the CI failure seen after the previous commit (it
did not reproduce locally, so this is not a confirmed cause), a module with no
imports of its own cannot participate in a cycle at all. The definition moves
there; transcript-index re-exports it so callers are unchanged.
